+++ src/main/java/com/getpcpanel/profile/Save.java
@@ -14,6 +14,7 @@ import lombok.extern.log4j.Log4j2;
 @Log4j2
 public class Save {
     private Map<String, DeviceSave> devices = new ConcurrentHashMap<>();
+    private boolean overlayEnabled;
     private boolean obsEnabled;
     private String obsAddress = "localhost";
     private String obsPort = "4455";