+++ src/main/java/com/getpcpanel/ui/SettingsDialog.java
@@ -41,6 +41,7 @@ public class SettingsDialog extends Application implements UIInitializer {
 
     private Stage stage;
     @FXML private Pane root;
+    @FXML public CheckBox overlay;
     @FXML private CheckBox obsEnable;
     @FXML private Pane obsControls;
     @FXML private TextField obsAddress;
@@ -113,6 +114,7 @@ public class SettingsDialog extends Application implements UIInitializer {
     @FXML
     private void ok(ActionEvent event) {
         var save = saveService.get();
+        save.setOverlayEnabled(overlay.isSelected());
         save.setObsEnabled(obsEnable.isSelected());
         save.setObsAddress(obsAddress.getText());
         save.setObsPort(obsPort.getText());
@@ -140,6 +142,7 @@ public class SettingsDialog extends Application implements UIInitializer {
 
     private void initFields() {
         var save = saveService.get();
+        overlay.setSelected(save.isOverlayEnabled());
         obsEnable.setSelected(save.isObsEnabled());
         obsAddress.setText(save.getObsAddress());
         obsPort.setText(save.getObsPort());